Add message sending functionality
The changes expose an endpoint that takes in a json with all the fields
in the database messages. The frontend is expected to have three different
ways to call this endpoint based on the type of message, leaving some
fields null as required, e.g. no width or height for a non-video message.